fix clang compile error

+++ contrib/dynamic_embedding/src/tde/details/multithreaded_id_transformer_benchmark.cpp
@@ -14,8 +14,12 @@ static void BM_MultiThreadedIDTransformer_Cold(benchmark::State& state) {
     global_ids.random_(1e10, 2e10);
     state.ResumeTiming();
     transformer.Transform(
-        tcb::span{global_ids.template data_ptr<int64_t>(), global_ids.numel()},
-        tcb::span{cache_ids.template data_ptr<int64_t>(), cache_ids.numel()});
+        tcb::span{
+            global_ids.template data_ptr<int64_t>(),
+            static_cast<size_t>(global_ids.numel())},
+        tcb::span{
+            cache_ids.template data_ptr<int64_t>(),
+            static_cast<size_t>(cache_ids.numel())});
   }
 }
 
@@ -36,8 +40,12 @@ static void BM_MultiThreadedIDTransformer_Hot(benchmark::State& state) {
     global_ids.random_(1e6, 2e6);
     state.ResumeTiming();
     transformer.Transform(
-        tcb::span{global_ids.template data_ptr<int64_t>(), global_ids.numel()},
-        tcb::span{cache_ids.template data_ptr<int64_t>(), cache_ids.numel()});
+        tcb::span{
+            global_ids.template data_ptr<int64_t>(),
+            static_cast<size_t>(global_ids.numel())},
+        tcb::span{
+            cache_ids.template data_ptr<int64_t>(),
+            static_cast<size_t>(cache_ids.numel())});
   }
 }
